# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vastra MVP API
  version: 1.1.0
  description: Canonical MVP API contract for Customer, Merchant, Captain, and Admin
    applications.
servers:
- url: http://localhost:8080/v1
  description: Local
- url: https://staging-api.vastra.in/v1
  description: Staging
- url: https://api.vastra.in/v1
  description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Forbidden:
      description: Actor is not authorized for the resource
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    NotFound:
      description: Resource not found or not visible to the actor
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Conflict:
      description: Resource state conflicts with the request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
...
    ServiceUnavailable:
      description: Service is temporarily unavailable
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
  schemas:
    ApiError:
      type: object
      required:
      - success
      - error
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
              - object
              - 'null'
              additionalProperties: true
            retryable:
              type: boolean
              default: false
        requestId:
          type:
          - string
          - 'null'
          format: uuid
...
    CustomerOrderAddress:
      type: object
      required: [id, label, recipientName, phoneNumber, line1, line2, landmark, area, city, state, postalCode, countryCode, latitude, longitude]
      properties:
        id: {type: string, format: uuid}
        label: {type: ['string', 'null']}
        recipientName: {type: string, minLength: 1}
        phoneNumber: {type: string, minLength: 1}
        line1: {type: string, minLength: 1}
        line2: {type: ['string', 'null']}
        landmark: {type: ['string', 'null']}
        area: {type: string, minLength: 1}
        city: {type: string, minLength: 1}
        state: {type: string, minLength: 1}
        postalCode: {type: string, minLength: 1}
        countryCode: {type: string, pattern: '^[A-Z]{2}$'}
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
    CustomerOrderShop:
      type: object
      required: [id, name, slug]
      properties:
        id: {type: string, format: uuid}
        name: {type: string, minLength: 1}
        slug: {type: string, minLength: 1}
    CustomerOrderItem:
      type: object
      required: [id, productId, variantId, productName, sku, colourName, sizeLabel, imageObjectKey, quantity, unitMrpPaise, unitSellingPricePaise, discountPaise, totalPaise]
      properties:
        id: {type: string, format: uuid}
        productId: {type: string, format: uuid}
        variantId: {type: string, format: uuid}
        productName: {type: string, minLength: 1}
        sku: {type: string, minLength: 1}
        colourName: {type: ['string', 'null']}
        sizeLabel: {type: ['string', 'null']}
        imageObjectKey: {type: ['string', 'null']}
        quantity: {type: integer, minimum: 1}
        unitMrpPaise: {type: integer, minimum: 0}
        unitSellingPricePaise: {type: integer, minimum: 0}
        discountPaise: {type: integer, minimum: 0}
        totalPaise: {type: integer, minimum: 0}
    CustomerOrderTotals:
      type: object
      required: [subtotalPaise, productDiscountPaise, couponDiscountPaise, deliveryFeePaise, platformFeePaise, taxPaise, totalPaise]
      properties:
        subtotalPaise: {type: integer, minimum: 0}
        productDiscountPaise: {type: integer, minimum: 0}
        couponDiscountPaise: {type: integer, minimum: 0}
        deliveryFeePaise: {type: integer, minimum: 0}
        platformFeePaise: {type: integer, minimum: 0}
        taxPaise: {type: integer, minimum: 0}
        totalPaise: {type: integer, minimum: 0}
    OrderStatus:
      type: string
      enum:
      - PAYMENT_PENDING
      - WAITING_FOR_MERCHANT
      - MERCHANT_ACCEPTED
      - PACKING
      - READY_FOR_PICKUP
      - CAPTAIN_SEARCHING
      - CAPTAIN_ASSIGNED
      - CAPTAIN_AT_STORE
      - PICKED_UP
      - OUT_FOR_DELIVERY
      - CAPTAIN_AT_CUSTOMER
      - DELIVERED
      - COMPLETED
      - PROBLEM_REPORTED
      - CANCELLED
    Order:
      type: object
      required: [id, orderNumber, cartId, quoteId, shop, address, status, paymentStatus, paymentMethod, fulfilmentType, items, totals, estimatedDeliveryAt, customerNote, placedAt, replayed]
      properties:
        id: {type: string, format: uuid}
        orderNumber: {type: string, minLength: 1}
        cartId: {type: string, format: uuid}
        quoteId: {type: string, format: uuid}
        shop: {$ref: '#/components/schemas/CustomerOrderShop'}
        address: {$ref: '#/components/schemas/CustomerOrderAddress'}
        status:
          type: string
          const: WAITING_FOR_MERCHANT
        paymentStatus:
          type: string
          const: COD_PENDING
        paymentMethod:
          type: string
          const: COD
        fulfilmentType:
          type: string
          const: DELIVERY
        items:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/CustomerOrderItem'}
        totals: {$ref: '#/components/schemas/CustomerOrderTotals'}
        estimatedDeliveryAt: {type: string, format: date-time}
        customerNote: {type: ['string', 'null'], maxLength: 500}
        placedAt: {type: string, format: date-time}
        replayed: {type: boolean}
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [cartId, quoteId, addressId, paymentMethod]
      properties:
        cartId: {type: string, format: uuid}
        quoteId: {type: string, format: uuid}
        addressId: {type: string, format: uuid}
        paymentMethod:
          type: string
          const: COD
          description: Online payment is outside the first order vertical slice.
        customerNote: {type: ['string', 'null'], maxLength: 500}
    PlaceCustomerOrderResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data:
          type: object
          required: [order]
          properties:
            order: {$ref: '#/components/schemas/Order'}
        meta:
          type: object
          required: [requestId]
          properties:
            requestId: {type: ['string', 'null'], format: uuid}
    CustomerOrderSummary:
      type: object
      required: [id, orderNumber, shop, status, paymentStatus, fulfilmentType, itemCount, previewImageObjectKey, totals, estimatedDeliveryAt, placedAt, createdAt]
      properties:
        id: {type: string, format: uuid}
        orderNumber: {type: string, minLength: 1}
        shop: {$ref: '#/components/schemas/CustomerOrderShop'}
        status: {$ref: '#/components/schemas/OrderStatus'}
        paymentStatus:
          type: string
          enum: [PENDING, AUTHORIZED, CAPTURED, FAILED, PARTIALLY_REFUNDED, REFUNDED, COD_PENDING, COD_COLLECTED]
        fulfilmentType:
          type: string
          enum: [DELIVERY, CUSTOMER_PICKUP]
        itemCount: {type: integer, minimum: 1}
        previewImageObjectKey: {type: ['string', 'null']}
        totals: {$ref: '#/components/schemas/CustomerOrderTotals'}
        estimatedDeliveryAt: {type: ['string', 'null'], format: date-time}
        placedAt: {type: ['string', 'null'], format: date-time}
        createdAt: {type: string, format: date-time}
...
    ListCustomerOrdersResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data:
          type: object
          required: [orders, nextCursor]
          properties:
            orders:
              type: array
              items: {$ref: '#/components/schemas/CustomerOrderSummary'}
            nextCursor: {type: ['string', 'null']}
        meta:
          type: object
          required: [requestId]
          properties:
            requestId: {type: ['string', 'null'], format: uuid}
...
    AdminOperationalOrder:
      type: object
      additionalProperties: false
      required: [id, orderNumber, status, paymentStatus, fulfilmentType, totalPaise, itemCount, customer, shop, deliveryTaskId, deliveryStatus, interventionReason, estimatedDeliveryAt, placedAt, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        orderNumber: {type: string}
        status:
          type: string
          enum: [PAYMENT_PENDING, WAITING_FOR_MERCHANT, MERCHANT_ACCEPTED, PACKING, READY_FOR_PICKUP, CAPTAIN_SEARCHING, CAPTAIN_ASSIGNED, CAPTAIN_AT_STORE, PICKED_UP, OUT_FOR_DELIVERY, CAPTAIN_AT_CUSTOMER, DELIVERED, COMPLETED, PROBLEM_REPORTED, CANCELLED]
        paymentStatus: {type: string}
        fulfilmentType: {type: string}
        totalPaise: {type: integer, minimum: 0}
        itemCount: {type: integer, minimum: 0}
        customer:
          type: object
          additionalProperties: false
          required: [id, name, phoneNumber]
          properties:
            id: {type: string, format: uuid}
            name: {type: ['string', 'null']}
            phoneNumber: {type: ['string', 'null']}
        shop:
          type: object
          additionalProperties: false
          required: [id, name, merchantId]
          properties:
            id: {type: string, format: uuid}
            name: {type: string}
            merchantId: {type: string, format: uuid}
        deliveryTaskId: {type: ['string', 'null'], format: uuid}
        deliveryStatus: {type: ['string', 'null']}
        interventionReason:
          type: ['string', 'null']
          enum: [DELAYED, UNASSIGNED, MERCHANT_TIMEOUT, CAPTAIN_ISSUE, PAYMENT_ISSUE, null]
        estimatedDeliveryAt: {type: ['string', 'null'], format: date-time}
        placedAt: {type: ['string', 'null'], format: date-time}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AdminOrderListResponse:
      type: object
      additionalProperties: false
      required: [orders, nextCursor]
      properties:
        orders:
          type: array
          items: {$ref: '#/components/schemas/AdminOperationalOrder'}
        nextCursor: {type: ['string', 'null']}
...
security:
- bearerAuth: []
paths:
...
  /orders:
    get:
      operationId: listCustomerOrders
      summary: List the authenticated customer's orders
      description: Returns newest orders first with immutable totals, shop snapshots, item counts, and an opaque continuation cursor. Results are constrained by authenticated order RLS.
      tags: [Customer]
      security: [{bearerAuth: []}]
      parameters:
      - name: cursor
        in: query
        required: false
        schema: {type: string}
      - name: limit
        in: query
        required: false
        schema: {type: integer, minimum: 1, maximum: 50, default: 20}
      responses:
        '200':
          description: Customer order summaries
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListCustomerOrdersResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500':
          description: Customer order summary data is internally inconsistent
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
        '503':
          description: Customer order history is temporarily unavailable
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
    post:
      operationId: placeCustomerCodOrder
      summary: Place one idempotent customer COD order
      description: Atomically validates the checkout quote against current cart, address, shop, price, and stock state; reserves inventory; creates immutable order snapshots and lifecycle history; converts the cart; and writes transactional outbox events. Online payment is not implemented in this slice.
      tags: [Customer]
      security: [{bearerAuth: []}]
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateOrderRequest'}
      responses:
        '200':
          description: COD order created or an identical request safely replayed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlaceCustomerOrderResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500':
          description: Order, quote, inventory, or lifecycle state is internally inconsistent
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
        '503':
          description: Order placement is temporarily unavailable
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
...
  /admin/dashboard:
    get:
      summary: Get admin dashboard
      tags:
      - Admin
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/orders:
    get:
      operationId: listAdminOperationalOrders
      summary: List operational orders
      description: Keyset-paginated operations queue with optional lifecycle and intervention filters.
      tags: [Admin]
      security: [{bearerAuth: []}]
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum: [PAYMENT_PENDING, WAITING_FOR_MERCHANT, MERCHANT_ACCEPTED, PACKING, READY_FOR_PICKUP, CAPTAIN_SEARCHING, CAPTAIN_ASSIGNED, CAPTAIN_AT_STORE, PICKED_UP, OUT_FOR_DELIVERY, CAPTAIN_AT_CUSTOMER, DELIVERED, COMPLETED, PROBLEM_REPORTED, CANCELLED]
      - name: issue
        in: query
        required: false
        schema:
          type: string
          enum: [DELAYED, UNASSIGNED, MERCHANT_TIMEOUT, CAPTAIN_ISSUE, PAYMENT_ISSUE]
      - name: cursor
        in: query
        required: false
        schema: {type: string}
      - name: limit
        in: query
        required: false
        schema: {type: integer, minimum: 1, maximum: 50, default: 20}
      responses:
        '200':
          description: Operational order page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminOrderListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...